+++ src/Circle.cpp
@@ -1,6 +1,6 @@
 #include "Circle.h"
 
-Circle::Circle(int x, int y, int r,int partitions,Operators3D * opGlobal):Exam2::Polygon(opGlobal,init(x,y,r,partitions)){
+Circle::Circle(GLfloat x, GLfloat y, GLfloat r,int partitions,Operators3D * opGlobal):Exam2::Polygon(opGlobal,init(x,y,r,partitions)){
 
 }
 
@@ -9,7 +9,7 @@ Circle::~Circle()
     //dtor
 }
 
-std::vector<Point> Circle::init(int x,int y,int r,int partitions){
+std::vector<Point> Circle::init(GLfloat x, GLfloat y, GLfloat r,int partitions){
     auto difRad = pi2/partitions;
     std::vector<Point> ans;
     for(GLfloat theta = 0; theta < pi2 ; theta += difRad){
